bump submodule to `main`, not `master`
I updated the submodule by running

    git submodule update --remote

thinking that it would update to latest.  Unfortunately due to the
weird way submodules work, the branch to pull wasn't specified
anywhere, so it defaulted to `master`, which we don't use any more.

This changes the .gitmodules configuration so that this command will
pull `main` now, and bumps the submodule to the correct commit.